safety static cast

+++ src/cpu/cpu_arm.cpp
@@ -570,7 +570,7 @@ namespace gbaemu
                            : ((inst >> 16) & 0x0F);
         uint16_t rList = thumb ? (inst & 0x0FF) : (inst & 0x0FFFF);
         if (patchRList) {
-            rList |= 1 << regs::PC_OFFSET;
+            rList |= static_cast<uint16_t>(1) << regs::PC_OFFSET;
         }
 
         auto currentRegs = state.getCurrentRegs();
